Add KineticsFamily.getRulesForTemplate and extractSourceFromComments functions
These functions try to trace back the original template,
rate rule, or families used to estimate the kinetics
based on the kinetics comments in a reaction.
For estimates from rate rules, it also provides the exact weights from the
geometric averaging.

To be used for tracking uncertainties, or other interesting
things with tracking rate rule sources