ThermoDatabase.extractSourceFromComments: parser for extracting thermo source from species object
To be used for uncertainty analysis and for
analyzing chemkin files of old models.

Parses the verbose string of comments from the thermo data of the species object,
and extracts the thermo sources.

Returns a dictionary with keys of either 'Library', 'QM', and/or 'GAV'.
Commonly, species thermo are estimated using only one of these sources.
However, a radical can be estimated with more than one type of source, for
instance a saturated library value and a GAV HBI correction, or a QM saturated value
and a GAV HBI correction.

source = {'Library': String_Name_of_Library_Used,
          'QM': String_of_Method_Used,
          'GAV': Dictionary_of_Groups_Used
          }

The Dictionary_of_Groups_Used is a a dictionary that has keys: groupType i.e.
groupType could be one of
['group','gauche','int15','other','ring','polycyclic', 'radical']